+++ packages/lumi-survey/tsup.config.ts
@@ -5,11 +5,7 @@ export default defineConfig({
   format: ["esm"],
   dts: true,
   clean: true,
-  external: [
-    "@navikt/nav-dekoratoren-moduler",
-    "@navikt/ds-react",
-    "@navikt/aksel-icons",
-  ],
+  external: ["@navikt/ds-react", "@navikt/aksel-icons"],
   loader: {
     ".module.css": "local-css",
     ".css": "css",